DAC is fully functional! Also changed longs to ints because Tucker thinks that 4 = 8.
git-svn-id: https://fusion.wpi.edu/svn/rbe3001a15team2/trunk@57 015edbb1-928d-476b-afdc-d2170ecc5f68

+++ RBELib/include/RBELib/pot.h
@@ -21,7 +21,7 @@ typedef struct
 
 typedef struct
 {
-	long min;
+	int min;
 	int adcChannel;
 	double scaler;
 } potentiometer;
